refactor: eliminate circular import dependencies
